Add destroy dependencies
Why:
Associations between the user, list and task
models exist, so when deleting one object it
is necessary to make sure that all its
dependencies are deleted from the db too.

This change addresses the need by:
* Adding `dependent: :destroy` to the user
model, next to the `has_many :lists`
statement.
* Adding `dependent: :destroy` to the list
model, next to the `has_many :tasks`
statement.