Link list_id to new task automatically
Why:
When creating a new task the user should not be
asked to enter the list_id for the task since he
before pressed `new task` within the list he
wanted it.

This change addresses the need by:
* Adding `new_task_path(@list.id)` to the `New Task` link in the
lists/show view to pass the id from the current_id.
* Adding in the tasks/_form view this
<%= form.hidden_field :list_id, :value => params[:format] %> so
that the list_id is hidden and it automatically takes the passed
parameter corresponding to the list_id.