Add extra attributes for User model
Why:
The instructions require the User to have attributes name and
description, so it is needed to add them since the devise
generation does not add those fields.

This change addresses the need by:
* Running the command `rails generate migration AddDetailsToUsers
name:string description:text` and `rails db:migrate` to add the
new columns to the db.
* Adding the necessary fields in the new and edit layouts so that
the they can be entered.
* Adding the keys for the new attributes (`:name` and
`:description`) within the listed permitted parameters in the
_application_controller_