Set mailer feature
Why?
The TODO App require to send weekly emails to all its users with
the uncompleted tasks from their lists.

This change addresses the need by:
* Generating a mailer in which to set the email account to be used
and define the mailer to call the mail(to:...) operation.
* Create the html and text views to structure the content for the
email body.
* Add the configurations for using an email account and Gmail in
the development environment.
* Installing Redis and setting its configurations, and adding the
gems `redis-rails`, `sidekiq` and `sidekiq-scheduler` to the
gemfile.
* Generate a sidekiq worker and call the mailer method defined
before in the `perform` action of the worker.
* Add the sidekiq configuration to perform every week.