Correct bug when no name, photo or description
Why?
The user should be able to have a proper view or layout even if he does
not entered a name, description or photo.

This change addresses the need by:
Adding the proper if conditions to the application.html so if the user
does not have those attributes the interface will require to display the
photo, name nor description.